Added Full Map Interaction Functionality
Now map can be loaded into
Camera now works with full support for clickability
For ease of use added a scene changer. Just press "B" to load Battlefield and "M" to load map

+++ Assets/Scripts/Persistence/Runtime/DataStorage/Files/Crusade/Records/CrusadeState.cs
@@ -1,12 +1,11 @@
 
-using Scellecs.Morpeh;
 using System.Collections.Generic;
 
 namespace Persistence.DS
 {
     public struct CrusadeState : IDataStorageRecord
     {
-        public int curr_node_id;
+        public byte curr_node_id;
 
         public STAGES curr_stage;
